fdf presque fini

+++ src/bresenham.c
@@ -6,7 +6,7 @@
 /*   By: lbesnard <lbesnard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/30 21:54:24 by lbesnard          #+#    #+#             */
-/*   Updated: 2022/04/08 00:05:32 by lbesnard         ###   ########.fr       */
+/*   Updated: 2022/04/08 18:22:55 by lbesnard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include <stdio.h>
 
 
-void	bresenham1(t_data img, t_point2d p1, t_point2d p2, float color)
+void	bresenham1(t_data img, t_point2d p1, t_point2d p2, int color)
 {
 	t_point2d	d;
 	int			e;
@@ -38,7 +38,7 @@ void	bresenham1(t_data img, t_point2d p1, t_point2d p2, float color)
 	}
 }
 
-void	bresenham2(t_data img, t_point2d p1, t_point2d p2, float color)
+void	bresenham2(t_data img, t_point2d p1, t_point2d p2, int color)
 {
 	t_point2d	d;
 	int			e;
@@ -61,7 +61,7 @@ void	bresenham2(t_data img, t_point2d p1, t_point2d p2, float color)
 	}
 }
 
-void	bresenham3(t_data img, t_point2d p1, t_point2d p2, float color)
+void	bresenham3(t_data img, t_point2d p1, t_point2d p2, int color)
 {
 	t_point2d	d;
 	int			e;
@@ -84,7 +84,7 @@ void	bresenham3(t_data img, t_point2d p1, t_point2d p2, float color)
 	}
 }
 
-void	bresenham4(t_data img, t_point2d p1, t_point2d p2, float color)
+void	bresenham4(t_data img, t_point2d p1, t_point2d p2, int color)
 {
 	t_point2d	d;
 	int			e;
